pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "requests>=2.32.3",
    "whois>=1.20240129.2",
]
